handle deps improvements

+++ handle_deps.sh
@@ -1,11 +1,20 @@
 #!/usr/bin/env bash
 CURRENT_DIR=`pwd`
 DEPS_DIR=${CURRENT_DIR}/dependencies
-
+USE_BOOST="no"
 CROW_DIR="crow"
 SPDLOG_DIR="spdlog"
 BOOST_DIR="boost"
 
+while getopts b: option
+do
+case "${option}" in
+b) USE_BOOST=${OPTARG};;
+esac
+done
+
+echo "IF BOOST IS REQUIRED ::  use ::  -b yes "
+echo "USE BOOST :: ${USE_BOOST}"
 
 if [ ! -d ${DEPS_DIR} ]; then
     mkdir -p ${DEPS_DIR}
@@ -37,16 +46,20 @@ fi
 # (this MIGHT work for Linux and MacOS :D )
 # https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.gz
 # READ:: http://www.boost.org/doc/libs/1_66_0/more/getting_started/unix-variants.html
-if [ ! -d ${BOOST_DIR} ]; then
-    wget https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.gz -O boost_install.tgz
-    tar -zxf boost_install.tgz
-    rm -f boost_install.tgz
-    ln -s boost_1_66_0/ ${BOOST_DIR}
 
-    #build boost
-    cd ${BOOST_DIR}
-    ./bootstrap.sh
-    ./b2 --build-dir=/tmp/build-boost toolset=gcc stage
+if [ "${USE_BOOST}" == "yes" ]; then
+    echo "CHECKING IF BOOST IS REQUIRED"
+    if [ ! -d ${BOOST_DIR} ]; then
+        wget https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.gz -O boost_install.tgz
+        tar -zxf boost_install.tgz
+        rm -f boost_install.tgz
+        ln -s boost_1_66_0/ ${BOOST_DIR}
+
+        #build boost
+        cd ${BOOST_DIR}
+        ./bootstrap.sh --prefix=/tmp/build-boost
+        ./b2 --build-dir=/tmp/build-boost toolset=gcc stage
+    fi
 fi
 
 #return to the initial place.
